Implement polly-style retries on requests and selenium gets
In order to increase reliability, I've made use of Retry and func_timeout to implement a resilience system. With this, failed web requests will be repeated up to a maximum amount of retries.

With this update, new unit tests have been implemented to test the retry mechanism.

I've also split testing requirements from prodution requirements in order to reduce the Heroku slug size